+++ demos/maptiler-sdk.umd.js
@@ -2644,6 +2644,9 @@
 	class Map$1 extends maplibregl.Map {
 	  constructor(options) {
 	    var _a;
+	    if (options.apiKey) {
+	      config.apiKey = options.apiKey;
+	    }
 	    const style = styleToStyle(options.style);
 	    const hashPreConstructor = location.hash;
 	    if (!config.apiKey) {
@@ -2734,8 +2737,11 @@
 	    this.on("styledataloading", () => {
 	      this.languageShouldUpdate = !!config.primaryLanguage || !!config.secondaryLanguage;
 	    });
+	    let initLanguageFromConstructor = true;
 	    this.on("styledata", () => {
-	      if (config.primaryLanguage && (this.languageShouldUpdate || !this.isStyleInitialized)) {
+	      if (options.language && initLanguageFromConstructor) {
+	        this.setPrimaryLanguage(options.language);
+	      } else if (config.primaryLanguage && (this.languageShouldUpdate || !this.isStyleInitialized)) {
 	        this.setPrimaryLanguage(config.primaryLanguage);
 	      }
 	      if (config.secondaryLanguage && (this.languageShouldUpdate || !this.isStyleInitialized)) {
@@ -2743,6 +2749,7 @@
 	      }
 	      this.languageShouldUpdate = false;
 	      this.isStyleInitialized = true;
+	      initLanguageFromConstructor = false;
 	    });
 	    this.on("styledata", () => {
 	      if (this.getTerrain() === null && this.isTerrainEnabled) {
@@ -2843,7 +2850,6 @@
 	      if (language === Language.AUTO) {
 	        return this.setPrimaryLanguage(getBrowserLanguage());
 	      }
-	      config.primaryLanguage = language;
 	      const layers = this.getStyle().layers;
 	      const strLanguageRegex = /^\s*{\s*name\s*(:\s*(\S*))?\s*}$/;
 	      const strLanguageInArrayRegex = /^\s*name\s*(:\s*(\S*))?\s*$/;
@@ -2917,7 +2923,6 @@
 	      if (language === Language.AUTO) {
 	        return this.setSecondaryLanguage(getBrowserLanguage());
 	      }
-	      config.secondaryLanguage = language;
 	      const layers = this.getStyle().layers;
 	      const strLanguageRegex = /^\s*{\s*name\s*(:\s*(\S*))?\s*}$/;
 	      const strLanguageInArrayRegex = /^\s*name\s*(:\s*(\S*))?\s*$/;
@@ -3061,6 +3066,12 @@
 	    hashBin[3] = this.getPitch();
 	    hashBin[4] = this.getBearing();
 	    return gBase64.fromUint8Array(new Uint8Array(hashBin.buffer));
+	  }
+	  getSdkConfig() {
+	    return config;
+	  }
+	  getMaptilerSessionId() {
+	    return MAPTILER_SESSION_ID;
 	  }
 	}